timer-list: Implement heap based timer-list
Previous timer-list was naive implementation of priority queue and very
slow when number of timers increased. This was not a problem because
only few timers were used. But with removal of dpd timer and replacement
with per-connection timer this may become problematic.

Solution is to use binary heap based priority queue which is much
faster.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>