qdevice: Add option to change socket file umask
Both qnetd and qdevice were not changing umask so local socket file
was usually not accessible by user from same user group as
qdevice/qnetd. This is mainly problem if daemon is running as coroqnetd
user and it is needed to run corosync-qnetd-tool as non-root user.

Patch adds advanced option (local_socket_umask) which allows to change
umask before socket is created.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>