qdevice: Add option to change socket file gid
Both qnetd and qdevice are not changing gid so local socket file
was usually not accessible by user if daemon is running as a root.

This is mainly problem if it is needed to run corosync-qdevice-tool as
non-root user.

Patch adds advanced option (local_socket_gid) which allows to change
group ownership after socket is created.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>